requirements.txt
pandas
numpy
alice-blue
kafka-python
pre-commit
pandas
numpy
alice-blue
pytest
